Add reviewers and commenters to commit message
This change makes the commit message more complete. Reviewers
are added in a "Reviewed-by:" tag, and commenters in a "Cc:"
one, as the Linux Kernel Guidelines states.

By doing this we can know more about the PR directly from
the git log, without having to go to the PR page.